notify: create test utils
